+++ src/main.cc
@@ -4,6 +4,7 @@ Function JSON::GetClass(Napi::Env env) {
   return DefineClass(env, "JSON",
                      {
                          JSON::InstanceAccessor<&JSON::TypeGetter>("type"),
+                         JSON::InstanceAccessor<&JSON::ToStringGetter>(Symbol::WellKnown(env, "toStringTag")),
                          JSON::InstanceMethod<&JSON::Get>("get"),
                          JSON::InstanceMethod<&JSON::Expand>("expand"),
                          JSON::InstanceMethod<&JSON::Path>("path"),